Move the ISO-related functions into their own subfolder.
