+++ include/benzina/benzina.h
@@ -15,7 +15,7 @@
 #include "benzina/config.h"
 #include "benzina/endian.h"
 #include "benzina/intops.h"
-#include "benzina/iso.h"
+#include "benzina/iso/iso.h"
 #include "benzina/version.h"
 #include "benzina/visibility.h"
