+++ include/benzina/iso/meson.build
@@ -1,0 +1,2 @@
+install_headers('iso.h',
+                install_dir: includedir/'benzina'/'iso')
