+++ src/libbenzina/iso/meson.build
@@ -1,0 +1,7 @@
+#
+# libbenzina ISO standard-related sources
+#
+
+libbenzina_dict += {
+    'srcs': libbenzina_dict['srcs'] + files('iso.c')
+}
